Delete a function that saves only the card name
The functions saves only the name without hash values.

This has an real use case. The user might want to reserve sequental
ZK numbers on the timeline. The reservation would have either NULL
or the SHA256 of an empty input.

This would case a special case in the database or it would have to be
handled separately in the code.

Too troublesome.